Make custom hook to check if the window is scrolled.
