--- src/components/navbar/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from "./navbar.helper";
